SPrintF: Allow strings to be given as arguments for numeric conversions
If a string is given as a parameter for a format specifier which expects a numeric type, convert the string to the numeric type first.

e.g.
* %c -> Take the first character of the string (don't convert to a number)
* %d -> Convert the string to a long integer, then format it
* %u -> Convert the string to a long unsigned integer, then format it
* %f, %d -> Convert the string to a double, then format it

The benefits are that strings can be automatically converted and formatted according to the format specifier, instead of just a string.

Issue: HELIOS-1585