SPrintF: Convert strings to number units using invariant culture
The feature 0a5ca4e ("SPrintF: Allow strings to be given as arguments for numeric conversions") converted the string to the numeric using the current locale. This works for locales which are also similar to the invariant (e.g. US English), but other number systems would misinterpret the string values (e.g. 3.14 on Romanian would be interpreted as 314).

The invariant culture is used, as when printing, the invariant culture is for the output, so it should also be used for conversions as the parameters.

Issue: HELIOS-1615