Ensure NuGet package generated doesn't depend on RJCP.MSBuildTasks
Need to indicate that the assets are all private to force .NET to not add the package as a dependency. According to MSDN docs, this shouldn't be needed as the default is already contentFiles;build;analyzers, but even if the RJCP.MSBuildTasks NuGet package was modified to remove the 'library' folder, it still didn't work (and building tRJCP.MSBuildTasks resulted in a new warning with an incorrect package structure).

Issue: DOTNET-367