SPrintF: Add test cases for specific binary representations of double
When determining how to calculate the multiplication lookup table based on the exponent, it is really important that every entry in that table is tested. To do that, we must provide exact bit representations of the double values and check that the conversion to a string is correct.

Issue: DOTNET-825